fix(mission-runtime): relocate pure resolve_mid8 into mission_runtime.identity
Move resolve_mid8/mid8_from_slug (pure, regex-only) down into the
mission_runtime layer so resolution.py no longer reaches up into
specify_cli.lanes for them; branch_naming.py re-exports both names
verbatim and the now-unneeded lanes allow-row is removed from
_MISSION_RUNTIME_ALLOWED_SPECIFY_CLI.
